requirements.txt
# Core NLP and text processing
spacy>=3.7.0
pydantic>=2.0.0
click>=8.0.0

# Data and utilities  
pyyaml>=6.0.0
python-dateutil>=2.8.0

# Testing and development
pytest>=7.0.0
pytest-cov>=4.0.0

# Future dependencies (commented for now)
# nltk>=3.8.0
# transformers>=4.30.0  # For ML-based claim detection
# requests>=2.31.0     # For fact-checking APIs
